removed deprecated methods from manual gating from the la module
